requirements.txt
alembic==1.7.7
amqp==5.1.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.3
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.1
six==1.16.0
SQLAlchemy==1.4.34
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.0
wrapt==1.14.0
zipp==3.7.0
python-dotenv==0.20.0
